Change cmake build in README to out-of-source
This is kind of a dumb thing to submit a PR for, but it caught my eye. The current `cmake` instructions in the README result in an in-source build. In-source builds are considered to be bad practice, and bad advice to give beginners. I changed the instructions to instead create an out-of-source build.

There are a number of reasons to prefer out-of-source over in-source, but the big one is that the resulting `build` dir is easy to nuke and recreate, no matter how badly you screw things up.