Remove TODO for passing argv
We don't need to manually pass argv as it will be set to sys.argv in
Traitlets (see https://github.com/ipython/traitlets/blob/11e840600ad740984034d3418142ec979b3ebfae/traitlets/config/application.py#L672).

We just need to be sure that sys.argv is correctly set, which is the
case for xpython, but not for the xpython wheel.